requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
idna==3.2
interrogate==1.5.0
lxml==4.6.5
openpyxl==3.0.9
prompt-toolkit==3.0.20
py==1.10.0
py7zr==0.16.2
questionary==1.10.0
pyodbc==4.0.32
python-docx==0.8.11
questionary==1.10.0
requests==2.26.0
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.7
wcwidth==0.2.5
